requirements.txt
streamlit==1.37.1
chromadb==0.5.4
sentence-transformers==3.0.1
transformers==4.43.3
torch>=2.2.0; platform_system != 'Darwin' or platform_machine != 'arm64'
pygit2==1.14.1
python-dotenv==1.0.1
tqdm==4.66.4
unidecode==1.3.8
numpy==1.26.4
pandas==2.2.2
pygments==2.18.0
rank-bm25==0.2.2
rapidfuzz==3.10.0
ragas==0.1.13
datasets==2.20.0
requests==2.32.3
groq==0.11.0

